requirements.txt
gradio>=4.44.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
reportlab>=4.2.0
Pillow>=10.0.0
